move time hack from actor_base to scada
